% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{data_gen}
\alias{data_gen}
\title{Function that returns values from optional seasonal deterministic model}
\usage{
data_gen(
  volatility,
  init_EIR = 100,
  min_EIR = 0.01,
  max_EIR = 500,
  EIR_step = 30,
  sample_size = 220,
  sample_sd = 40,
  out_step = 30,
  time = 5 * 365,
  prop_treated = 0.4,
  init_age = c(0, 0.25, 0.5, 0.75, 1, 1.25, 1.5, 1.75, 2, 3.5, 5, 7.5, 10, 15, 20, 30,
    40, 50, 60, 70, 80),
  plot_instance = TRUE
)
}
\arguments{
\item{volatility}{Volatility of the random walk}

\item{init_EIR}{Initial EIR value to kick of the random walk}

\item{min_EIR}{Minimum value over which the random walk will not exceed.}

\item{max_EIR}{Maximum value over which the random walk will not exceed.}

\item{EIR_step}{This controls the frequency with which EIR changes in the model.
This is the length of time (in days) for each EIR period.}

\item{sample_size}{Average sample size for number of tested individuals in the simulated data set.}

\item{sample_sd}{Standard deviation of the normal distribution of
sample size for the number of tested individuals in the simulated data set.}

\item{out_step}{The number of days between each simulated data point.}

\item{time}{The length of time in days the model will run.}

\item{prop_treated}{Proportion of the cases that are treated}

\item{init_age}{Age groups within the model}

\item{plot_instance}{If TRUE, plots the simulated prevalence.}
}
\description{
\code{data_gen} Generate simulated data based on a random walk of EIR values
}
