% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_piecewise.R
\name{data_gen_piecewise}
\alias{data_gen_piecewise}
\title{Function that returns values from deterministic model based on user-specified series of EIR values}
\usage{
data_gen_piecewise(
  EIR_vals,
  EIR_times,
  sample_size = 220,
  sample_sd = 40,
  out_step = 30,
  duration = 5 * 365,
  prop_treated = 0.4,
  init_age = c(0, 0.25, 0.5, 0.75, 1, 1.25, 1.5, 1.75, 2, 3.5, 5, 7.5, 10, 15, 20, 30,
    40, 50, 60, 70, 80),
  plot_instance = TRUE
)
}
\arguments{
\item{EIR_vals}{Initial EIR value to kick of the random walk}

\item{EIR_times}{Minimum value over which the random walk will not exceed.}

\item{sample_size}{Average sample size for number of tested individuals in the simulated data set.}

\item{sample_sd}{Standard deviation of the normal distribution of
sample size for the number of tested individuals in the simulated data set.}

\item{out_step}{The number of days between each simulated data point.}

\item{duration}{The length of time in days the model will run.}

\item{prop_treated}{Proportion of the cases that are treated}

\item{init_age}{Age groups within the model}

\item{plot_instance}{If TRUE, plots the simulated prevalence.}
}
\description{
\code{data_gen_piecewise} Generate simulated data based on a user-specified time series of EIR values
}
