% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_seas.R
\name{data_gen_seas}
\alias{data_gen_seas}
\title{Function that returns values from optional seasonal deterministic model}
\usage{
data_gen_seas(
  init_EIR = 100,
  out_step = 30,
  sample_size = 220,
  sample_sd = 40,
  duration = 10 * 365,
  init_age = c(0, 0.25, 0.5, 0.75, 1, 1.25, 1.5, 1.75, 2, 3.5, 5, 7.5, 10, 15, 20, 30,
    40, 50, 60, 70, 80),
  prop_treated = 0.4,
  country = "Burkina Faso",
  admin_unit = "Cascades",
  plot_instance = TRUE
)
}
\arguments{
\item{init_EIR}{Initial EIR value to kick of the random walk}

\item{out_step}{The number of days between each simulated data point.}

\item{sample_size}{Average sample size for number of tested individuals in the simulated data set.}

\item{sample_sd}{Standard deviation of the normal distribution of
sample size for the number of tested individuals in the simulated data set.}

\item{duration}{The length of time in days the model will run.}

\item{init_age}{Age groups within the model}

\item{prop_treated}{Proportion of the cases that are treated}

\item{country}{Country to be used for seasonal profile}

\item{admin_unit}{Admin unit to be used for seasonal profile}

\item{plot_instance}{If TRUE, plots the simulated prevalence.}
}
\description{
\code{data_gen_seas} Generate simulated data based on a random walk of EIR values
}
