% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genrandwalk.R
\name{genrandwalk}
\alias{genrandwalk}
\title{Function that returns values from optional seasonal deterministic model}
\usage{
genrandwalk(x, vol, randWalk, min_EIR, max_EIR)
}
\arguments{
\item{x}{position in list}

\item{vol}{Volatility of the random walk}

\item{randWalk}{Values in random walk so far}

\item{min_EIR}{Minimum EIR value allowed}

\item{max_EIR}{Maximum EIR value allowed}
}
\description{
\code{genrandwalk} generate random walk of EIR (recursive fn)
}
